Refactor CLI into package modules
