+++ prompt_optimizer/__main__.py
@@ -2,7 +2,7 @@
 Main entry point for the DSPy Prompt Optimizer package.
 """
 
-from .cli import main
+from prompt_optimizer.cli import main
 
 if __name__ == "__main__":
     main()
